+++ src/ewokssphinx/tests/test_task_directive.py
@@ -1,53 +1,7 @@
 import pytest
 from sphinx.testing import restructuredtext
 
-from .conftest import assert_task_nodes
-
-
-def _assert_class_task_nodes(nodes):
-    assert_task_nodes(
-        nodes,
-        identifier="ewokssphinx.tests.dummy_tasks.MyTask",
-        doc="""My task documentation""",
-        task_type="class",
-        required_inputs=["a", "b", "c"],
-        optional_inputs=["d", "e"],
-        outputs=["error", "result"],
-    )
-
-
-def _assert_method_nodes(nodes):
-    assert_task_nodes(
-        nodes[0],
-        identifier="ewokssphinx.tests.dummy_tasks.my_other_task",
-        doc=None,
-        task_type="method",
-        required_inputs=["f", "g"],
-        optional_inputs=["h"],
-        outputs=["return_value"],
-    )
-
-    assert_task_nodes(
-        nodes[1],
-        identifier="ewokssphinx.tests.dummy_tasks.run",
-        doc="""Run task documentation""",
-        task_type="method",
-        required_inputs=["i"],
-        optional_inputs=["j", "k"],
-        outputs=["return_value"],
-    )
-
-
-def _assert_ppfmethod_nodes(nodes):
-    assert_task_nodes(
-        nodes,
-        identifier="ewokssphinx.tests.dummy_tasks.run",
-        doc="""Run task documentation""",
-        task_type="ppfmethod",
-        required_inputs=["i"],
-        optional_inputs=["j", "k"],
-        outputs=["return_value"],
-    )
+from .test_utils import assert_task_nodes
 
 
 def test_ewokstasks(app):
@@ -115,3 +69,49 @@ def test_ewokstasks_ignore_import_error(app):
     )
 
     assert len(parsed_nodes) == 0
+
+
+def _assert_class_task_nodes(nodes):
+    assert_task_nodes(
+        nodes,
+        identifier="ewokssphinx.tests.dummy_tasks.MyTask",
+        doc="""My task documentation""",
+        task_type="class",
+        required_inputs=["a", "b", "c"],
+        optional_inputs=["d", "e"],
+        outputs=["error", "result"],
+    )
+
+
+def _assert_method_nodes(nodes):
+    assert_task_nodes(
+        nodes[0],
+        identifier="ewokssphinx.tests.dummy_tasks.my_other_task",
+        doc=None,
+        task_type="method",
+        required_inputs=["f", "g"],
+        optional_inputs=["h"],
+        outputs=["return_value"],
+    )
+
+    assert_task_nodes(
+        nodes[1],
+        identifier="ewokssphinx.tests.dummy_tasks.run",
+        doc="""Run task documentation""",
+        task_type="method",
+        required_inputs=["i"],
+        optional_inputs=["j", "k"],
+        outputs=["return_value"],
+    )
+
+
+def _assert_ppfmethod_nodes(nodes):
+    assert_task_nodes(
+        nodes,
+        identifier="ewokssphinx.tests.dummy_tasks.run",
+        doc="""Run task documentation""",
+        task_type="ppfmethod",
+        required_inputs=["i"],
+        optional_inputs=["j", "k"],
+        outputs=["return_value"],
+    )